+++ ui/zenoedit/dock/docktabcontent.h
@@ -66,19 +66,6 @@ protected:
     static const int sToolbarHeight;
 };
 
-struct DockContentWidgetInfo {
-    int resolutionX;
-    int resolutionY;
-    bool lock;
-    int comboboxindex;
-    double colorR;
-    double colorG;
-    double colorB;
-    DockContentWidgetInfo(int resX, int resY, bool block, int index, double r, double g, double b) : resolutionX(resX), resolutionY(resY), lock(block), comboboxindex(index), colorR(r), colorG(g), colorB(b) {}
-    DockContentWidgetInfo(int resX, int resY, bool block, int index) : resolutionX(resX), resolutionY(resY), lock(block), comboboxindex(index) {}
-    DockContentWidgetInfo() : resolutionX(0), resolutionY(0), lock(false), comboboxindex(0), colorR(0.18f), colorG(0.20f), colorB(0.22f) {}
-};
-
 class DockContent_Parameter : public DockToolbarWidget
 {
     Q_OBJECT
