+++ ui/zenoio/reader/zsgreader.cpp
@@ -8,7 +8,6 @@
 #include "common.h"
 #include <zenomodel/customui/customuirw.h>
 #include "iotags.h"
-#include <zenoedit/layout/winlayoutrw.h>
 
 using namespace zeno::iotags;
 using namespace zeno::iotags::curve;
@@ -358,7 +357,7 @@ void ZsgReader::_parseSettings(const rapidjson::Value& jsonSettings, IAcceptor* 
     if (jsonSettings.HasMember("layoutinfo") && jsonSettings["layoutinfo"].IsObject())
     {
         LAYOUT_SETTING layout;
-        layout.layerOutNode = readLayout(jsonSettings["layoutinfo"]);
+        layout.layerOutNode = _readLayout(jsonSettings["layoutinfo"]);
         pAcceptor->setLayoutInfo(layout);
     }
 }
@@ -894,4 +893,67 @@ bool ZsgReader::_parseParams2(const QString& id, const QString &nodeCls, const r
     return true;
 }
 
+PtrLayoutNode ZsgReader::_readLayout(const rapidjson::Value& objValue)
+{
+    if (objValue.HasMember("orientation") && objValue.HasMember("left") && objValue.HasMember("right"))
+    {
+        PtrLayoutNode ptrNode = std::make_shared<LayerOutNode>();
+        QString ori = objValue["orientation"].GetString();
+        ptrNode->type = (ori == "H" ? NT_HOR : NT_VERT);
+        ptrNode->pLeft = _readLayout(objValue["left"]);
+        ptrNode->pRight = _readLayout(objValue["right"]);
+        ptrNode->pWidget = nullptr;
+        return ptrNode;
+    }
+    else if (objValue.HasMember("widget"))
+    {
+        PtrLayoutNode ptrNode = std::make_shared<LayerOutNode>();
+        ptrNode->type = NT_ELEM;
+        ptrNode->pLeft = nullptr;
+        ptrNode->pRight = nullptr;
+
+        const rapidjson::Value& widObj = objValue["widget"];
+
+        auto tabsObj = widObj["tabs"].GetArray();
+        QStringList tabs;
+        for (int i = 0; i < tabsObj.Size(); i++)
+        {
+            if (tabsObj[i].IsString())
+            {
+                ptrNode->tabs.push_back(tabsObj[i].GetString());
+            }
+            else if (tabsObj[i].IsObject())
+            {
+                if (tabsObj[i].HasMember("type") && QString(tabsObj[i]["type"].GetString()) == "View")
+                {
+                    ptrNode->tabs.push_back("View");
+                    DockContentWidgetInfo info(tabsObj[i]["resolutionX"].GetInt(), tabsObj[i]["resolutionY"].GetInt(),
+                        tabsObj[i]["blockwindow"].GetBool(), tabsObj[i]["resolution-combobox-index"].GetInt(), tabsObj[i]["backgroundcolor"][0].GetDouble(),
+                        tabsObj[i]["backgroundcolor"][1].GetDouble(), tabsObj[i]["backgroundcolor"][2].GetDouble());
+                    ptrNode->widgetInfos.push_back(info);
+                }
+                else if (tabsObj[i].HasMember("type") && QString(tabsObj[i]["type"].GetString()) == "Optix")
+                {
+                    ptrNode->tabs.push_back("Optix");
+                    DockContentWidgetInfo info(tabsObj[i]["resolutionX"].GetInt(), tabsObj[i]["resolutionY"].GetInt(),
+                        tabsObj[i]["blockwindow"].GetBool(), tabsObj[i]["resolution-combobox-index"].GetInt());
+                    ptrNode->widgetInfos.push_back(info);
+                }
+            }
+        }
+
+        const rapidjson::Value& geomObj = widObj["geometry"];
+        float x = geomObj["x"].GetFloat();
+        float y = geomObj["y"].GetFloat();
+        float width = geomObj["width"].GetFloat();
+        float height = geomObj["height"].GetFloat();
+        ptrNode->geom = QRectF(x, y, width, height);
+
+        return ptrNode;
+    }
+    else
+    {
+        return nullptr;
+    }
+}
