+++ ui/zenoio/reader/zsgreader.h
@@ -18,6 +18,7 @@ public:
     static ZsgReader& getInstance();
     bool openFile(const QString& fn, IAcceptor* pAcceptor);
     bool importNodes(IGraphsModel* pModel, const QModelIndex& subgIdx, const QString& nodeJson, const QPointF& targetPos, IAcceptor* pAcceptor);
+    PtrLayoutNode _readLayout(const rapidjson::Value& objValue);
 
 private:
     ZsgReader();
