+++ ui/zenoio/writer/zsgwriter.cpp
@@ -6,7 +6,6 @@
 #include "variantptr.h"
 #include <zenomodel/include/viewparammodel.h>
 #include <zenomodel/customui/customuirw.h>
-#include <zenoedit/layout/winlayoutrw.h>
 
 using namespace zeno::iotags;
 
@@ -539,7 +538,7 @@ void ZsgWriter::dumpSettings(const APP_SETTINGS settings, RAPIDJSON_WRITER& writ
         writer.EndObject();
 
         writer.Key("layoutinfo");
-        writeLayout(settings.layoutInfo.layerOutNode, settings.layoutInfo.size, writer);
+        _writeLayout(settings.layoutInfo.layerOutNode, settings.layoutInfo.size, writer, settings.layoutInfo.cbDumpTabsToZsg);
     }
 }
 
@@ -591,6 +590,75 @@ void ZsgWriter::_dumpDescriptors(const NODE_DESCS& descs, RAPIDJSON_WRITER& writ
 
             writer.Key("categories");
             AddStringList(desc.categories, writer);
+        }
+    }
+}
+
+void ZsgWriter::_writeLayout(PtrLayoutNode root, const QSize& szMainwin, PRETTY_WRITER& writer, void(*cbDumpTabsToZsg)(QDockWidget*, RAPIDJSON_WRITER&))
+{
+    JsonObjBatch scope(writer);
+    if (root->type == NT_HOR || root->type == NT_VERT)
+    {
+        writer.Key("orientation");
+        writer.String(root->type == NT_HOR ? "H" : "V");
+        writer.Key("left");
+        if (root->pLeft)
+            _writeLayout(root->pLeft, szMainwin, writer, cbDumpTabsToZsg);
+        else
+            writer.Null();
+
+        writer.Key("right");
+        if (root->pRight)
+            _writeLayout(root->pRight, szMainwin, writer, cbDumpTabsToZsg);
+        else
+            writer.Null();
+    }
+    else
+    {
+        writer.Key("widget");
+        if (root->pWidget == nullptr || root->pWidget->isHidden())
+        {
+            writer.Null();
+        }
+        else
+        {
+            writer.StartObject();
+            int w = szMainwin.width();
+            int h = szMainwin.height();
+            if (w == 0)
+                w = 1;
+            if (h == 0)
+                h = 1;
+
+            writer.Key("geometry");
+            writer.StartObject();
+            QRect rc = root->pWidget->geometry();
+
+            writer.Key("x");
+            float _left = (float)rc.left() / w;
+            writer.Double(_left);
+
+            writer.Key("y");
+            float _top = (float)rc.top() / h;
+            writer.Double(_top);
+
+            writer.Key("width");
+            float _width = (float)rc.width() / w;
+            writer.Double(_width);
+
+            writer.Key("height");
+            float _height = (float)rc.height() / h;
+            writer.Double(_height);
+
+            writer.EndObject();
+
+            writer.Key("tabs");
+            writer.StartArray();
+            if (cbDumpTabsToZsg)
+                cbDumpTabsToZsg(root->pWidget, writer);
+            writer.EndArray();
+
+            writer.EndObject();
         }
     }
 }
