+++ ui/zenoio/writer/zsgwriter.h
@@ -16,6 +16,7 @@ public:
     static ZsgWriter& getInstance();
     QString dumpProgramStr(IGraphsModel* pModel, APP_SETTINGS settings);
     void dumpToClipboard(const QMap<QString, NODE_DATA>& nodes);
+    void _writeLayout(PtrLayoutNode root, const QSize& szMainwin, PRETTY_WRITER& writer, void(*cbDumpTabsToZsg)(QDockWidget*, RAPIDJSON_WRITER&));
 
 private:
     ZsgWriter();
